added training folders and videos to .gitignore
